requirements.txt
streamlit
pandas
openpyxl
langchain
langchain-community
langchain-openai
langsmith
pydantic
python-dotenv
fpdf
python-docx
pyPDF2
docx2txt
unstructured
openai
python-doctr[tf]
Pillow
tensorflow
chromadb
opencv-python-headless
